crete MarkupInjector, fix logger
