+++ src/scripts/BaseView.js
@@ -5,7 +5,6 @@ import Logger from './Logger.js';
 export default class BaseView {
     imgUrl;
     id;
-    type;
     html;
 
     constructor(src, id) {
@@ -19,29 +18,9 @@ export default class BaseView {
      */
     _create(idParent) {
         if ( !document.getElementById(idParent) ) {
-            Logger.logCreateViewError('Не найден родительский элемент');
+            Logger.logError('Не найден родительский элемент');
         }
         
         document.getElementById(idParent).insertAdjacentHTML('beforeend', this.html);        
     }
-
-    /** 
-     * Create wrapper for elements, insert wrapper to document
-     * @param {string} title - wrapper title. 
-     * @param {string} id - wrapper id. 
-     * @param {string} idParent -  parent element id for wrapper. 
-     */
-    static createWrapper(title, id, idParent) {
-        const  html = `
-        <div class="content-item">
-            <div class="content-item__header">
-                <div class="content-item__title"><span>${title}</span>
-            </div>
-            </div>
-        
-            <div class="content-item__cards" id="${id}"></div>`;
-
-        document.getElementById(idParent).insertAdjacentHTML('beforeend', html);
-    }
-
 }
